Split code into smaller modules
