+++ crates/relayer-components/src/logger/traits/has_logger.rs
@@ -1,0 +1,10 @@
+use crate::core::traits::sync::Async;
+use crate::logger::traits::logger::BaseLogger;
+
+pub trait HasLoggerType: Async {
+    type Logger: BaseLogger;
+}
+
+pub trait HasLogger: HasLoggerType {
+    fn logger(&self) -> &Self::Logger;
+}
