+++ crates/relayer-components/src/logger/traits/mod.rs
@@ -1,1 +1,5 @@
-pub mod types;
+pub mod has_logger;
+pub mod logger;
+pub mod simple;
+pub mod value;
+pub mod wrapper;
