+++ crates/relayer-runtime/src/tokio/logger/level.rs
@@ -1,0 +1,14 @@
+#[derive(Clone)]
+pub enum LogLevel {
+    Trace,
+    Debug,
+    Info,
+    Warn,
+    Error,
+}
+
+impl Default for LogLevel {
+    fn default() -> Self {
+        Self::Info
+    }
+}
